Add Mode Select Cluster
** Problem **

Add the implementation for Mode Select Cluster. The specification for
Mode Select Cluster can be found at https://github.com/CHIP-Specifications/connectedhomeip-spec/blob/master/src/app_clusters/ModeSelect.adoc

** Change overview **

* Add src/app/zap-templates/zcl/data-model/chip/mode-select-cluster.xml
* Add the ModeSelectCluster to the all-cluster-app
* Add the ModeSelectCluster to the chip-tool
* Add src/app/clusters/mode-select-server/mode-select-server.cpp, the
main logic for the server of the ModeSelectCluster
* Add src/app/clusters/mode-select-server/supported-modes-manager.h,
which contains helper method for managing the supported options.
* Generate code for Mode Select Cluster

** Testing **
* Tested with chip-tool

+++ src/app/clusters/operational-credentials-server/operational-credentials-server.cpp
@@ -273,7 +273,7 @@ class OpCredsFabricTableDelegate : public FabricTableDelegate
     // Gets called when a fabric is deleted from KVS store
     void OnFabricDeletedFromStorage(FabricIndex fabricId) override
     {
-        emberAfPrintln(EMBER_AF_PRINT_DEBUG, "OpCreds: Fabric 0x%" PRIX16 " was deleted from fabric storage.", fabricId);
+        emberAfPrintln(EMBER_AF_PRINT_DEBUG, "OpCreds: Fabric 0x%" PRIX16 " was deleted from fabric storage.", (unsigned short) fabricId);
         writeFabricsIntoFabricsListAttribute();
     }
 
@@ -283,7 +283,7 @@ class OpCredsFabricTableDelegate : public FabricTableDelegate
         emberAfPrintln(EMBER_AF_PRINT_DEBUG,
                        "OpCreds: Fabric 0x%" PRIX16 " was retrieved from storage. FabricId 0x" ChipLogFormatX64
                        ", NodeId 0x" ChipLogFormatX64 ", VendorId 0x%04" PRIX16,
-                       fabric->GetFabricIndex(), ChipLogValueX64(fabric->GetFabricId()),
+                       (unsigned short) fabric->GetFabricIndex(), ChipLogValueX64(fabric->GetFabricId()),
                        ChipLogValueX64(fabric->GetPeerId().GetNodeId()), fabric->GetVendorId());
         writeFabricsIntoFabricsListAttribute();
     }
